@@ -1,6 +1,6 @@
 import Foundation
 
-public protocol DeviceCenterRepositoryProtocol: RepositoryProtocol {
+public protocol DeviceCenterRepositoryProtocol: RepositoryProtocol, Sendable {
     func fetchUserDevices() async -> [DeviceEntity]
     func fetchDeviceNames() async -> [String]
     func loadCurrentDeviceId() -> String?
