@@ -6,7 +6,7 @@ import SwiftUI
 import XCTest
 
 final class DeviceListViewRouterTests: XCTestCase {
-    
+    @MainActor
     func testBuild_rendersCorrectViewController() throws {
         let (sut, _) = try makeSUT()
         
@@ -15,6 +15,7 @@ final class DeviceListViewRouterTests: XCTestCase {
         XCTAssert(resultViewController is UIHostingController<DeviceListView>)
     }
     
+    @MainActor
     func testStart_pushCorrectViewController() throws {
         let (sut, mockPresenter) = try makeSUT()
         
@@ -26,9 +27,8 @@ final class DeviceListViewRouterTests: XCTestCase {
     }
     
     // MARK: - Helpers
-    
+    @MainActor
     private func makeSUT() throws -> (sut: DeviceListViewRouter, mockPresenter: UINavigationController) {
-        
         let mockPresenter = UINavigationController()
         let deviceCenterUseCase = MockDeviceCenterUseCase()
         let cameraUploadsUseCase = MockCameraUploadsUseCase(
